fixed sliders so you can slide on top or below also

+++ include/scene.h
@@ -20,7 +20,7 @@ public:
     Scene(Settings& settings, GLFWwindow* window, InputScheme* p_input_scheme, MasterRenderer& masterRenderer);
     ~Scene();
     virtual void render() = 0;
-    virtual void scene_logic(float delta_time) = 0;
+    virtual void scene_logic(double delta_time) = 0;
     InputScheme* p_input_scheme;
     MasterRenderer& masterRenderer;
     Camera camera;
@@ -35,12 +35,12 @@ class FreeCamWorld: public Scene
 public:
     FreeCamWorld(Settings& settings, GLFWwindow* window, MasterRenderer& masterRenderer);
     void render();
-    void scene_logic(float delta_time);
+    void scene_logic(double delta_time);
     World world;
     UI_FreeCamWorld ui;
 
-    void slider_amp_fun(float val, float change);
-    void slider_size_fun(float val, float change);
+    void slider_amp_fun(double val, double change);
+    void slider_size_fun(double val, double change);
 protected:
 private:
 };
@@ -51,8 +51,8 @@ class MainMenu: public Scene
 public:
     MainMenu(Settings& settings, GLFWwindow* window, MasterRenderer& masterRenderer);
     void render();
-    void scene_logic(float delta_time);
-    void test(float val, float change);
+    void scene_logic(double delta_time);
+    void test(double val, double change);
     UI_MainMenu ui;
 protected:
 private:
